+++ NEWS.md
@@ -1,3 +1,9 @@
+# sircovid 0.3.8
+
+- 4th model added to package - serology2_model
+- PMCMC now supports piecewise linear beta
+- Fixed bug in serology_model where there was some double counting in cum_new_conf
+
 # sircovid 0.3.7
 
 - Introduced a piecewise linear generate beta function